refact: preference -> bookPrefernce 변경 및 가독성 향상
