+++ src/main/java/com/ureca/idle/originapi/business/bookPreference/BookPreferenceManagingUseCase.java
@@ -1,0 +1,7 @@
+package com.ureca.idle.originapi.business.bookPreference;
+
+public interface BookPreferenceManagingUseCase {
+
+    void processLike(Long bookId, Long kidId);
+    void processDislike(Long bookId, Long kidId);
+}
