--- src/main/java/com/ureca/idle/originapi/business/preference/PreferenceManagingUseCase.java
@@ -1,7 +1,0 @@
-package com.ureca.idle.originapi.business.preference;
-
-public interface PreferenceManagingUseCase {
-
-    void processLike(Long bookId, Long kidId);
-    void processDislike(Long bookId, Long kidId);
-}
